serve/inferenceserver/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Redis dependencies
redis>=5.0.0
hiredis>=2.0.0
rq>=1.15.0

# Earth2Studio dependencies (if available)
# earth2studio[dlwp] @ git+https://github.com/NVIDIA/earth2studio.git

# skip this - PyTorch (for device management)
# torch>=2.0.0


# Metrics
prometheus_client>=0.17.0

# Additional utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.0.0
hydra-core>=1.3.0


# Cryptography for CloudFront signed URLs
cryptography>=41.0.0

# Multi-Storage Client with Azure support
multi-storage-client[azure-storage-blob]
# Azure Blob Storage for SAS token generation
azure-storage-blob>=12.19.0
azure-identity

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
